+++ CHANGES.md
@@ -1,6 +1,6 @@
 IFCtoRDF Change Log
 
 Release 0.1 (2017-05-26)
+ - change to buildingsmart-tech namespace
  - Use maven-java-formatter-plugin to format all code
- - ISSUE-3 Use maven-java-formatter-plugin to format all code
  - Mavenize codebase